+++ packages/snap/tests/e2e/importStateBackup.spec.ts
@@ -1,11 +1,15 @@
-import { type MascaState } from '@blockchain-lab-um/masca-types';
+import {
+  CURRENT_STATE_VERSION,
+  MascaLegacyStateV1,
+  type MascaState,
+} from '@blockchain-lab-um/masca-types';
 import { Result, isError } from '@blockchain-lab-um/utils';
 import { IDataManagerSaveResult } from '@blockchain-lab-um/veramo-datamanager';
 import { MetaMaskInpageProvider } from '@metamask/providers';
 import { SnapsProvider } from '@metamask/snaps-sdk';
 import type { IIdentifier, VerifiableCredential } from '@veramo/core';
 import cloneDeep from 'lodash.clonedeep';
-import { beforeAll, describe, expect, it } from 'vitest';
+import { beforeAll, describe, expect, it, vi } from 'vitest';
 
 import { onRpcRequest } from '../../src';
 import StorageService from '../../src/storage/Storage.service';
@@ -15,6 +19,16 @@ import { EXAMPLE_VC_PAYLOAD } from '../data/credentials';
 import { getDefaultSnapState } from '../data/defaultSnapState';
 import { createTestVCs } from '../helpers/generateTestVCs';
 import { SnapMock, createMockSnap } from '../helpers/snapMock';
+import {
+  getEmptyAccountState,
+  getInitialSnapState,
+} from '../../src/utils/config';
+import EncryptionService from '../../src/Encryption.service';
+import {
+  getLegacyEmptyAccountStateV1,
+  getLegacyStateV1,
+} from '../data/legacyStates/legacyStateV1';
+import { randomUUID } from 'crypto';
 
 describe('importStateBackup', () => {
   let snapMock: SnapsProvider & SnapMock;
@@ -60,9 +74,88 @@ describe('importStateBackup', () => {
     }
   });
 
+  it('Should suceed with v1 empty state', async () => {
+    const spy = vi.spyOn(StorageService, 'migrateState');
+
+    const legacyStateV1 = getLegacyStateV1();
+    legacyStateV1.v1.accountState[account] = getLegacyEmptyAccountStateV1();
+    legacyStateV1.v1.currentAccount = account;
+
+    const encryptedState = await EncryptionService.encrypt(
+      JSON.stringify(legacyStateV1)
+    );
+
+    const importStateBackupResult = (await onRpcRequest({
+      origin: 'http://localhost',
+      request: {
+        id: 'test-id',
+        jsonrpc: '2.0',
+        method: 'importStateBackup',
+        params: { serializedState: encryptedState },
+      },
+    })) as Result<unknown>;
+
+    if (isError(importStateBackupResult)) {
+      throw new Error(importStateBackupResult.error);
+    }
+
+    const expectedState = getInitialSnapState();
+    expectedState[CURRENT_STATE_VERSION].accountState[account] =
+      getEmptyAccountState();
+    expectedState[CURRENT_STATE_VERSION].currentAccount = account;
+
+    expect(spy).toHaveBeenCalled();
+    expect(StorageService.get()).toEqual(expectedState);
+    expect.assertions(2);
+  });
+
+  it('Should suceed with v1 non-empty state (1 credential)', async () => {
+    const spy = vi.spyOn(StorageService, 'migrateState');
+
+    const legacyStateV1 = getLegacyStateV1();
+    const credentialId = randomUUID();
+    legacyStateV1.v1.accountState[account] = getLegacyEmptyAccountStateV1();
+    legacyStateV1.v1.currentAccount = account;
+    legacyStateV1.v1.accountState[account].veramo.credentials = {
+      [credentialId]: generatedVC,
+    };
+
+    const encryptedState = await EncryptionService.encrypt(
+      JSON.stringify(legacyStateV1)
+    );
+
+    const importStateBackupResult = (await onRpcRequest({
+      origin: 'http://localhost',
+      request: {
+        id: 'test-id',
+        jsonrpc: '2.0',
+        method: 'importStateBackup',
+        params: { serializedState: encryptedState },
+      },
+    })) as Result<unknown>;
+
+    if (isError(importStateBackupResult)) {
+      throw new Error(importStateBackupResult.error);
+    }
+
+    const expectedState = getInitialSnapState();
+    expectedState[CURRENT_STATE_VERSION].accountState[account] =
+      getEmptyAccountState();
+    expectedState[CURRENT_STATE_VERSION].currentAccount = account;
+    expectedState[CURRENT_STATE_VERSION].accountState[
+      account
+    ].veramo.credentials = {
+      [credentialId]: generatedVC,
+    };
+
+    expect(spy).toHaveBeenCalled();
+    expect(StorageService.get()).toEqual(expectedState);
+    expect.assertions(2);
+  });
+
   it('Should suceed with default empty state', async () => {
     const startState: MascaState = cloneDeep(StorageService.get());
-    const resExport = (await onRpcRequest({
+    const exportStateBackupResult = (await onRpcRequest({
       origin: 'http://localhost',
       request: {
         id: 'test-id',
@@ -72,22 +165,22 @@ describe('importStateBackup', () => {
       },
     })) as Result<string>;
 
-    if (isError(resExport)) {
-      throw new Error(resExport.error);
+    if (isError(exportStateBackupResult)) {
+      throw new Error(exportStateBackupResult.error);
     }
 
-    const resImport = (await onRpcRequest({
+    const importStateBackupResult = (await onRpcRequest({
       origin: 'http://localhost',
       request: {
         id: 'test-id',
         jsonrpc: '2.0',
         method: 'importStateBackup',
-        params: { serializedState: resExport.data },
+        params: { serializedState: exportStateBackupResult.data },
       },
     })) as Result<unknown>;
 
-    if (isError(resImport)) {
-      throw new Error(resImport.error);
+    if (isError(importStateBackupResult)) {
+      throw new Error(importStateBackupResult.error);
     }
 
     expect(StorageService.get()).toEqual(startState);
@@ -95,7 +188,7 @@ describe('importStateBackup', () => {
   });
 
   it('Should suceed with non-empty state (1 credential)', async () => {
-    const saveRes = (await onRpcRequest({
+    const saveCredentialResult = (await onRpcRequest({
       origin: 'http://localhost',
       request: {
         id: 'test-id',
@@ -108,11 +201,11 @@ describe('importStateBackup', () => {
       },
     })) as Result<IDataManagerSaveResult[]>;
 
-    if (isError(saveRes)) {
-      throw new Error(saveRes.error);
+    if (isError(saveCredentialResult)) {
+      throw new Error(saveCredentialResult.error);
     }
 
-    expect(saveRes.data).toEqual([
+    expect(saveCredentialResult.data).toEqual([
       {
         id: expect.any(String),
         store: ['snap'],
@@ -120,7 +213,7 @@ describe('importStateBackup', () => {
     ]);
 
     const startState: MascaState = cloneDeep(StorageService.get());
-    const resExport = (await onRpcRequest({
+    const exportStateBackupResult = (await onRpcRequest({
       origin: 'http://localhost',
       request: {
         id: 'test-id',
@@ -130,22 +223,22 @@ describe('importStateBackup', () => {
       },
     })) as Result<string>;
 
-    if (isError(resExport)) {
-      throw new Error(resExport.error);
+    if (isError(exportStateBackupResult)) {
+      throw new Error(exportStateBackupResult.error);
     }
 
-    const resImport = (await onRpcRequest({
+    const importStateBackupResult = (await onRpcRequest({
       origin: 'http://localhost',
       request: {
         id: 'test-id',
         jsonrpc: '2.0',
         method: 'importStateBackup',
-        params: { serializedState: resExport.data },
+        params: { serializedState: exportStateBackupResult.data },
       },
     })) as Result<unknown>;
 
-    if (isError(resImport)) {
-      throw new Error(resImport.error);
+    if (isError(importStateBackupResult)) {
+      throw new Error(importStateBackupResult.error);
     }
     expect(StorageService.get()).toEqual(startState);
     expect.assertions(2);