+++ packages/snap/tests/unit/Storage.service.spec.ts
@@ -1,7 +1,4 @@
-import {
-  CURRENT_STATE_VERSION,
-  MascaLegacyStateV1,
-} from '@blockchain-lab-um/masca-types';
+import { CURRENT_STATE_VERSION } from '@blockchain-lab-um/masca-types';
 import { MetaMaskInpageProvider } from '@metamask/providers';
 import { SnapsProvider } from '@metamask/snaps-sdk';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
@@ -9,7 +6,7 @@ import { beforeEach, describe, expect, it, vi } from 'vitest';
 import StorageService from '../../src/storage/Storage.service';
 import { getInitialSnapState } from '../../src/utils/config';
 import { SnapMock, createMockSnap } from '../helpers/snapMock';
-import * as MigrateState from '../../src/utils/stateMigration';
+import { getLegacyStateV1 } from '../data/legacyStates/legacyStateV1';
 
 describe('Storage Service', () => {
   let snapMock: SnapsProvider & SnapMock;
@@ -62,61 +59,47 @@ describe('Storage Service', () => {
     );
     expect.assertions(2);
   });
-});
-describe('State Migration', () => {
-  let snapMock: SnapsProvider & SnapMock;
-  beforeEach(async () => {
-    snapMock = createMockSnap();
-    snapMock.rpcMocks.snap_manageState({
-      operation: 'clear',
-    });
-    global.snap = snapMock;
-    global.ethereum = snapMock as unknown as MetaMaskInpageProvider;
+
+  it('should migrate state from v1 to latest version', async () => {
+    const legacyStateV1 = getLegacyStateV1();
+
+    const newState = StorageService.migrateState(
+      structuredClone(legacyStateV1)
+    );
+
+    const expectedState = getInitialSnapState();
+
+    expect(newState).toEqual(expectedState);
+
+    expect.assertions(1);
   });
 
   it('should not migrate state from latest version', async () => {
-    const spy = vi.spyOn(MigrateState, 'migrateToV2');
     const state = getInitialSnapState();
-    StorageService.set(state);
 
-    await StorageService.save();
+    const newState = StorageService.migrateState(structuredClone(state));
 
-    const newState = StorageService.get();
+    expect(newState).toEqual(state);
 
-    expect(newState).toHaveProperty('v2');
-
-    StorageService.init();
-    await StorageService.save();
-
-    const newState2 = StorageService.get();
-    expect(newState2).toHaveProperty('v2');
-    expect(spy).toHaveBeenCalledTimes(0);
-
-    expect.assertions(3);
+    expect.assertions(1);
   });
 
-  it('should succeed migrating state from v1 to v2', async () => {
-    const spy = vi.spyOn(MigrateState, 'migrateToV2');
-    let state: any = getInitialSnapState();
-    state = { v1: state.v2 };
-    state.v1.config.dApp.friendlyDapps = ['masca.io'];
-    state.v1.config.dApp.permissions = undefined;
-    const legacyStateV1: MascaLegacyStateV1 = state;
-    StorageService.set(legacyStateV1 as any);
+  it('should migrate state to latest version when init method is called', async () => {
+    const spy = vi.spyOn(StorageService, 'migrateState');
 
+    const legacyStateV1 = getLegacyStateV1();
+    StorageService.set(legacyStateV1 as any);
     await StorageService.save();
 
-    const newState = StorageService.get();
-
-    expect(newState).toHaveProperty('v1');
+    let state = StorageService.get();
+    expect(state).toHaveProperty('v1');
 
     StorageService.init();
     await StorageService.save();
 
-    const newState2 = StorageService.get();
-    expect(newState2).toHaveProperty('v2');
-    expect(spy).toHaveBeenCalled();
-
+    state = StorageService.get();
+    expect(state).toHaveProperty(CURRENT_STATE_VERSION);
+    expect(spy).toHaveBeenCalledOnce();
     expect.assertions(3);
   });
 });