+++ libs/did-provider-key/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.0.8-beta.1
+
+### Patch Changes
+
+- [#589](https://github.com/blockchain-lab-um/masca/pull/589) [`96ba0ab`](https://github.com/blockchain-lab-um/masca/commit/96ba0ab2da593844a988e71316529e7f1c1e02ba) Thanks [@andyv09](https://github.com/andyv09)! - Build packages with Tsup
+
+- Updated dependencies [[`96ba0ab`](https://github.com/blockchain-lab-um/masca/commit/96ba0ab2da593844a988e71316529e7f1c1e02ba)]:
+  - @blockchain-lab-um/utils@1.3.8-beta.0
+
 ## 1.0.8-beta.0
 
 ### Patch Changes