+++ packages/dapp/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @blockchain-lab-um/dapp
 
+## 1.2.1-beta.1
+
+### Patch Changes
+
+- Updated dependencies [[`96ba0ab`](https://github.com/blockchain-lab-um/masca/commit/96ba0ab2da593844a988e71316529e7f1c1e02ba)]:
+  - @blockchain-lab-um/did-provider-key@1.0.8-beta.1
+  - @blockchain-lab-um/masca-connector@1.3.1-beta.1
+
 ## 1.2.1-beta.0
 
 ### Patch Changes